+++ tests/scripts/script_hook_process.py
@@ -1,0 +1,11 @@
+#
+# Author: Sébastien Helleu <flashcode@flashtux.org>
+#
+
+"""A WeeChat script."""
+
+import weechat
+
+if __name__ == "__main__":
+    if weechat.register("script", "author", "0.1", "GPL3", "desc", "", ""):
+        weechat.hook_process("url:http://localhost:1234", 1000, "callback", "")